requirements.txt
Django==4.2.16
django-nested-inline==0.4.6
Pillow==10.4.0
gunicorn==20.1.0
whitenoise==6.5.0
